{Fix} Nouveau Doublon Script Enemy
Fusion des deux Scripts Enemy pour ne pas avoir de doublon.

+++ Assets/Scripts/Fights/CanTakeDamage.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
 public interface ICanTakeDamage {
-    public void TakeDamage(float damage); //DEBUG!!! l'appliquer aux ennemis
+    public void TakeDamage(int damage); //DEBUG!!! l'appliquer aux ennemis
     public GameObject WhoAmI();
 }
